requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
#psycopg2-binary==2.9.9  # for later Postgres; not required if staying on SQLite
Werkzeug==3.0.1
itsdangerous==2.2.0
